requirement.txt
flask
flask-restplus
pyserial
psutil
wmi
flask_pymongo
flask_cors
iteration-utilities